--- src/channels/turn/kernel.test.ts
@@ -1,2300 +1,0 @@
-// Channel turn kernel tests cover orchestration, dispatch, and completion behavior.
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { ReplyPayload } from "../../auto-reply/reply-payload.js";
-import type { HistoryEntry } from "../../auto-reply/reply/history.types.js";
-import type { DispatchReplyWithBufferedBlockDispatcher } from "../../auto-reply/reply/provider-dispatcher.types.js";
-import type { FinalizedMsgContext } from "../../auto-reply/templating.js";
-import type { OpenClawConfig } from "../../config/types.openclaw.js";
-import {
-  emitTrustedDiagnosticEvent,
-  onInternalDiagnosticEvent,
-  resetDiagnosticEventsForTest,
-  waitForDiagnosticEventsDrained,
-  type DiagnosticEventPayload,
-} from "../../infra/diagnostic-events.js";
-import {
-  createChildDiagnosticTraceContext,
-  freezeDiagnosticTraceContext,
-  getActiveDiagnosticTraceContext,
-} from "../../infra/diagnostic-trace-context.js";
-import { PlatformMessageNotDispatchedError } from "../../infra/outbound/deliver-types.js";
-import { logMessageProcessed } from "../../logging/diagnostic.js";
-import { getChildLogger, resetLogger, setLoggerOverride } from "../../logging/logger.js";
-import { outboundMessageIdentities } from "../message/outbound-echo-state.js";
-import { recordOutboundMessageIdentity } from "../message/outbound-echo.js";
-import type { RecordInboundSession } from "../session.types.js";
-import {
-  hasFinalChannelTurnDispatch,
-  hasVisibleChannelTurnDispatch,
-  resolveChannelTurnDispatchCounts,
-} from "./dispatch-result.js";
-import { runPreparedChannelTurn } from "./execution.js";
-import { dispatchAssembledChannelTurn, dispatchRoutedChannelTurn } from "./lifecycle.js";
-import { runChannelTurn } from "./run-channel-turn.js";
-import type { ChannelTurnResult } from "./types.js";
-
-const deliverOutboundPayloads = vi.hoisted(() => vi.fn());
-const resolveOutboundDurableFinalDeliverySupport = vi.hoisted(() => vi.fn());
-const sendDurableMessageBatch = vi.hoisted(() => vi.fn());
-const recordInboundSessionCore = vi.hoisted(() => vi.fn(async () => undefined));
-const dispatchReplyWithBufferedBlockDispatcherCore = vi.hoisted(() => vi.fn());
-const dispatchReplyWithRoutedChannelDispatcherCore = vi.hoisted(() => vi.fn());
-const emitMessageSent = vi.hoisted(() => vi.fn());
-const getGlobalHookRunner = vi.hoisted(() => vi.fn());
-const createMessageSentEmitter = vi.hoisted(() =>
-  vi.fn(() => ({ emitMessageSent, hasMessageSentHooks: true })),
-);
-const readRecentUserAssistantTextForSession = vi.hoisted(() => vi.fn());
-
-vi.mock("../../auto-reply/reply/provider-dispatcher.js", async (importOriginal) => {
-  const actual =
-    await importOriginal<typeof import("../../auto-reply/reply/provider-dispatcher.js")>();
-  return {
-    ...actual,
-    dispatchReplyWithBufferedBlockDispatcher: dispatchReplyWithBufferedBlockDispatcherCore,
-  };
-});
-
-vi.mock("../../auto-reply/dispatch.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../../auto-reply/dispatch.js")>();
-  return {
-    ...actual,
-    dispatchInboundMessageWithRoutedChannelDispatcher: dispatchReplyWithRoutedChannelDispatcherCore,
-  };
-});
-
-vi.mock("../../infra/outbound/deliver.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../../infra/outbound/deliver.js")>();
-  return {
-    ...actual,
-    deliverOutboundPayloads,
-    resolveOutboundDurableFinalDeliverySupport,
-  };
-});
-
-vi.mock("../message/send.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../message/send.js")>();
-  return {
-    ...actual,
-    sendDurableMessageBatch,
-  };
-});
-
-vi.mock("../session.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../session.js")>();
-  return { ...actual, recordInboundSession: recordInboundSessionCore };
-});
-
-vi.mock("../../infra/outbound/message-sent-hook.js", () => ({
-  createMessageSentEmitter,
-}));
-
-vi.mock("../../plugins/hook-runner-global.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../../plugins/hook-runner-global.js")>();
-  return { ...actual, getGlobalHookRunner };
-});
-
-vi.mock("../../config/sessions/transcript.js", () => ({
-  readRecentUserAssistantTextForSession,
-}));
-
-const cfg = {} as OpenClawConfig;
-
-function createCtx(overrides: Partial<FinalizedMsgContext> = {}): FinalizedMsgContext {
-  return {
-    Body: "hello",
-    RawBody: "hello",
-    CommandBody: "hello",
-    From: "sender",
-    To: "target",
-    SessionKey: "agent:main:test:peer",
-    Provider: "test",
-    Surface: "test",
-    ...overrides,
-  } as FinalizedMsgContext;
-}
-
-function createRecordInboundSession(events: string[] = []): RecordInboundSession {
-  return vi.fn(async () => {
-    events.push("record");
-  }) as unknown as RecordInboundSession;
-}
-
-function createDispatch(
-  events: string[] = [],
-  deliverPayload: { text: string } = { text: "reply" },
-): DispatchReplyWithBufferedBlockDispatcher {
-  return vi.fn(async (params) => {
-    events.push("dispatch");
-    await params.dispatcherOptions.deliver(deliverPayload, { kind: "final" });
-    return {
-      queuedFinal: true,
-      counts: { tool: 0, block: 0, final: 1 },
-    };
-  }) as DispatchReplyWithBufferedBlockDispatcher;
-}
-
-function dispatchTestAssembledTurn(
-  overrides: Omit<
-    Parameters<typeof dispatchAssembledChannelTurn>[0],
-    "cfg" | "agentId" | "storePath"
-  >,
-) {
-  return dispatchAssembledChannelTurn({
-    cfg,
-    agentId: "main",
-    storePath: "/tmp/sessions.json",
-    ...overrides,
-  });
-}
-
-function requireFirstMockCall<T>(mock: { mock: { calls: T[][] } }, label: string): T[] {
-  const call = mock.mock.calls[0];
-  if (!call) {
-    throw new Error(`expected ${label} call`);
-  }
-  return call;
-}
-
-function createDurableSendResult(messageIds: string[]) {
-  return {
-    status: "sent",
-    results: messageIds.map((messageId) => ({ messageId })),
-    receipt: {
-      platformMessageIds: messageIds,
-      parts: [],
-      sentAt: 1,
-    },
-  };
-}
-
-type DurableSendRequest = {
-  accountId?: string;
-  channel?: string;
-  durability?: string;
-  payloads?: ReplyPayload[];
-  replyToMode?: string;
-  session?: {
-    key?: string;
-    agentId?: string;
-    requesterAccountId?: string;
-    requesterSenderId?: string;
-    conversationType?: string;
-  };
-  threadId?: string | number | null;
-  to?: string;
-};
-
-type DurableSupportRequest = {
-  channel?: string;
-  requirements?: Record<string, boolean>;
-};
-
-type DeliveryResult = {
-  messageIds?: string[];
-  receipt?: { platformMessageIds?: string[] };
-  visibleReplySent?: boolean;
-};
-
-type FinalizeResult = {
-  admission?: unknown;
-  dispatched?: boolean;
-  routeSessionKey?: string;
-};
-
-type TurnLogEvent = {
-  event?: string;
-  messageId?: string;
-  stage?: string;
-};
-
-function latestDurableSendRequest(): DurableSendRequest {
-  const calls = sendDurableMessageBatch.mock.calls;
-  const call = calls[calls.length - 1] as unknown as [DurableSendRequest] | undefined;
-  if (!call) {
-    throw new Error("expected durable send request");
-  }
-  const [request] = call;
-  return request;
-}
-
-function latestDurableSupportRequest(): DurableSupportRequest {
-  const calls = resolveOutboundDurableFinalDeliverySupport.mock.calls;
-  const call = calls[calls.length - 1] as unknown as [DurableSupportRequest] | undefined;
-  if (!call) {
-    throw new Error("expected durable support request");
-  }
-  const [request] = call;
-  return request;
-}
-
-function deliveryResult(value: unknown): DeliveryResult {
-  return value as DeliveryResult;
-}
-
-function finalizeResult(value: unknown): FinalizeResult {
-  return value as FinalizeResult;
-}
-
-function expectDispatched<TDispatchResult>(
-  result: ChannelTurnResult<TDispatchResult>,
-): asserts result is Extract<ChannelTurnResult<TDispatchResult>, { dispatched: true }> {
-  expect(result.dispatched).toBe(true);
-  if (!result.dispatched) {
-    throw new Error("expected dispatch");
-  }
-}
-
-function loggedEvents(log: ReturnType<typeof vi.fn>): TurnLogEvent[] {
-  return log.mock.calls.map(([event]) => {
-    const entry = event as TurnLogEvent;
-    return {
-      stage: entry.stage,
-      event: entry.event,
-      ...(entry.messageId === undefined ? {} : { messageId: entry.messageId }),
-    };
-  });
-}
-
-describe("channel turn kernel", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    recordInboundSessionCore.mockResolvedValue(undefined);
-    dispatchReplyWithBufferedBlockDispatcherCore.mockImplementation(createDispatch());
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementation(createDispatch());
-    outboundMessageIdentities.clear();
-    resetDiagnosticEventsForTest();
-    resetLogger();
-    setLoggerOverride({ level: "info" });
-    resolveOutboundDurableFinalDeliverySupport.mockResolvedValue({ ok: true });
-    createMessageSentEmitter.mockImplementation(() => ({
-      emitMessageSent,
-      hasMessageSentHooks: true,
-    }));
-    getGlobalHookRunner.mockReturnValue(null);
-    readRecentUserAssistantTextForSession.mockResolvedValue([]);
-  });
-
-  afterEach(() => {
-    setLoggerOverride(null);
-    resetLogger();
-  });
-
-  it("runs routed direct message hooks after payload preparation", async () => {
-    const events: string[] = [];
-    const runMessageSending = vi.fn(async (event: { content: string }) => {
-      events.push("message_sending");
-      return { content: `${event.content} + message-hook` };
-    });
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    const deliver = vi.fn(async (payload: ReplyPayload) => {
-      events.push("deliver");
-      return { messageIds: ["direct-1"], visibleReplySent: true, content: payload.text };
-    });
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      accountId: "acct",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({
-        Surface: "telegram",
-        OriginatingTo: "chat-1",
-        ReplyToId: "source-1",
-        MessageThreadId: 42,
-      }),
-      delivery: {
-        preparePayload: (payload) => {
-          events.push("prepare");
-          return { ...payload, text: `${payload.text} + prepared`, mediaUrls: ["media://1"] };
-        },
-        deliver,
-      },
-    });
-
-    expect(events).toEqual(["prepare", "message_sending", "deliver"]);
-    expect(deliver).toHaveBeenCalledWith(
-      { text: "reply + prepared + message-hook", mediaUrls: ["media://1"] },
-      { kind: "final" },
-    );
-    expect(runMessageSending).toHaveBeenCalledWith(
-      expect.objectContaining({
-        content: "reply + prepared",
-        replyToId: "source-1",
-        threadId: 42,
-        metadata: expect.objectContaining({
-          channel: "telegram",
-          accountId: "acct",
-          mediaUrls: ["media://1"],
-        }),
-      }),
-      expect.objectContaining({
-        channelId: "telegram",
-        accountId: "acct",
-        conversationId: "chat-1",
-        sessionKey: "agent:main:telegram:peer",
-      }),
-    );
-    expectDispatched(result);
-    expect(result.dispatchResult.counts.final).toBe(1);
-  });
-
-  it("does not let message hooks resurrect payloads suppressed during preparation", async () => {
-    const runMessageSending = vi.fn(async () => ({ content: "resurrected" }));
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    const durable = vi.fn();
-    const deliver = vi.fn();
-    const onDelivered = vi.fn();
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "whatsapp",
-      route: { agentId: "main", sessionKey: "agent:main:whatsapp:peer" },
-      ctxPayload: createCtx({ Surface: "whatsapp", OriginatingTo: "chat-1" }),
-      delivery: {
-        preparePayload: () => null,
-        durable,
-        deliver,
-        onDelivered,
-      },
-    });
-
-    expect(runMessageSending).not.toHaveBeenCalled();
-    expect(durable).not.toHaveBeenCalled();
-    expect(deliver).not.toHaveBeenCalled();
-    expect(onDelivered).toHaveBeenCalledWith(
-      { text: "reply" },
-      { kind: "final" },
-      {
-        visibleReplySent: false,
-        suppression: { reason: "no_visible_payload" },
-      },
-    );
-    expectDispatched(result);
-    expect(result.dispatchResult).toMatchObject({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-    });
-  });
-
-  it("suppresses routed direct delivery and visible counts when message hooks cancel", async () => {
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending: vi.fn(async () => ({
-        cancel: true,
-        cancelReason: "policy",
-        metadata: { source: "test" },
-      })),
-    });
-    const deliver = vi.fn();
-    const onDelivered = vi.fn();
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram", OriginatingTo: "chat-1" }),
-      delivery: { deliver, onDelivered, observeMessageSent: true },
-    });
-
-    expect(deliver).not.toHaveBeenCalled();
-    expect(onDelivered).toHaveBeenCalledWith(
-      { text: "reply" },
-      { kind: "final" },
-      {
-        visibleReplySent: false,
-        suppression: {
-          reason: "cancelled_by_message_sending_hook",
-          cancelReason: "policy",
-          metadata: { source: "test" },
-        },
-      },
-    );
-    expect(emitMessageSent).not.toHaveBeenCalled();
-    expectDispatched(result);
-    expect(result.dispatchResult).toMatchObject({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-    });
-    expect(hasVisibleChannelTurnDispatch(result.dispatchResult)).toBe(false);
-  });
-
-  it("exposes media-only routed payloads to the message hook before direct delivery", async () => {
-    const runMessageSending = vi.fn(async () => ({ cancel: true }));
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    const deliver = vi.fn();
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementationOnce(async (params) => {
-      await params.dispatcherOptions.deliver({ mediaUrls: ["media://only"] }, { kind: "final" });
-      return { queuedFinal: true, counts: { tool: 0, block: 0, final: 1 } };
-    });
-
-    await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram", OriginatingTo: "chat-1" }),
-      delivery: { deliver },
-    });
-
-    expect(runMessageSending).toHaveBeenCalledWith(
-      expect.objectContaining({
-        content: "",
-        metadata: expect.objectContaining({ mediaUrls: ["media://only"] }),
-      }),
-      expect.anything(),
-    );
-    expect(deliver).not.toHaveBeenCalled();
-  });
-
-  it("reconciles one cancelled payload without hiding a delivered sibling", async () => {
-    const runMessageSending = vi.fn(async ({ content }: { content: string }) =>
-      content === "cancel me" ? { cancel: true } : undefined,
-    );
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    const deliver = vi.fn(async () => ({ visibleReplySent: true }));
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementationOnce(async (params) => {
-      await params.dispatcherOptions.deliver({ text: "deliver me" }, { kind: "block" });
-      await params.dispatcherOptions.deliver({ text: "cancel me" }, { kind: "final" });
-      return { queuedFinal: true, counts: { tool: 0, block: 1, final: 1 } };
-    });
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram" }),
-      delivery: { deliver },
-    });
-
-    expect(runMessageSending).toHaveBeenCalledTimes(2);
-    expect(deliver).toHaveBeenCalledTimes(1);
-    expect(deliver).toHaveBeenCalledWith({ text: "deliver me" }, { kind: "block" });
-    expectDispatched(result);
-    expect(result.dispatchResult).toMatchObject({
-      queuedFinal: false,
-      counts: { tool: 0, block: 1, final: 0 },
-    });
-    expect(hasVisibleChannelTurnDispatch(result.dispatchResult)).toBe(true);
-  });
-
-  it("delegates routed hybrid delivery to the provider message hook owner", async () => {
-    const runMessageSending = vi.fn();
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    const deliverWithProviderMessageSending = vi.fn(async () => ({
-      messageIds: ["provider-1"],
-      visibleReplySent: true,
-    }));
-
-    await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram" }),
-      delivery: { deliverWithProviderMessageSending },
-    });
-
-    expect(deliverWithProviderMessageSending).toHaveBeenCalledWith(
-      { text: "reply" },
-      { kind: "final" },
-    );
-    expect(runMessageSending).not.toHaveBeenCalled();
-  });
-
-  it("uses the durable message hook owner and the core owner only after unsupported preflight", async () => {
-    const runMessageSending = vi.fn(async ({ content }: { content: string }) => ({
-      content: `${content} + direct-hook`,
-    }));
-    getGlobalHookRunner.mockReturnValue({
-      hasHooks: (name: string) => name === "message_sending",
-      runMessageSending,
-    });
-    sendDurableMessageBatch.mockResolvedValueOnce(createDurableSendResult(["durable-1"]));
-    const durableDeliver = vi.fn();
-
-    await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram", To: "chat-1" }),
-      delivery: { deliver: durableDeliver, durable: { replyToMode: "first" } },
-    });
-
-    expect(durableDeliver).not.toHaveBeenCalled();
-    expect(runMessageSending).not.toHaveBeenCalled();
-
-    resolveOutboundDurableFinalDeliverySupport.mockResolvedValueOnce({
-      ok: false,
-      reason: "missing_outbound_handler",
-    });
-    const fallbackDeliver = vi.fn(async () => ({ visibleReplySent: true }));
-    await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram", To: "chat-1" }),
-      delivery: { deliver: fallbackDeliver, durable: { replyToMode: "first" } },
-    });
-
-    expect(runMessageSending).toHaveBeenCalledTimes(1);
-    expect(fallbackDeliver).toHaveBeenCalledWith(
-      { text: "reply + direct-hook" },
-      { kind: "final" },
-    );
-  });
-
-  it("classifies routed delivery only after provider finalization settles", async () => {
-    const onDelivered = vi.fn();
-    const finalization = Promise.resolve({
-      messageIds: ["final-1"],
-      visibleReplySent: true as const,
-      content: "final content",
-    });
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram" }),
-      delivery: {
-        deliver: async () => ({ visibleReplySent: false, finalization }),
-        onDelivered,
-      },
-    });
-
-    expect(onDelivered).toHaveBeenCalledWith(
-      { text: "reply" },
-      { kind: "final" },
-      expect.objectContaining({
-        messageIds: ["final-1"],
-        visibleReplySent: true,
-        content: "final content",
-      }),
-    );
-    expectDispatched(result);
-    expect(result.dispatchResult.counts.final).toBe(1);
-    expect(result.dispatchResult.queuedFinal).toBe(true);
-  });
-
-  it("routes assembled final replies through durable outbound delivery", async () => {
-    sendDurableMessageBatch.mockResolvedValueOnce(createDurableSendResult(["tg-1"]));
-    const deliver = vi.fn();
-    const recordInboundSession = createRecordInboundSession();
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    const result = await dispatchTestAssembledTurn({
-      channel: "telegram",
-      accountId: "acct",
-      routeSessionKey: "agent:main:telegram:peer",
-      ctxPayload: createCtx({
-        To: "123",
-        OriginatingTo: "123",
-        MessageThreadId: 777,
-        AccountId: "acct",
-        ChatType: "group",
-        SenderId: "sender-1",
-      }),
-      recordInboundSession,
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver, durable: { replyToMode: "first" } },
-    });
-
-    expect(result.dispatched).toBe(true);
-    expect(deliver).not.toHaveBeenCalled();
-    expect(sendDurableMessageBatch).toHaveBeenCalledTimes(1);
-    const sendRequest = latestDurableSendRequest();
-    expect(sendRequest.channel).toBe("telegram");
-    expect(sendRequest.to).toBe("123");
-    expect(sendRequest.accountId).toBe("acct");
-    expect(sendRequest.payloads?.[0]?.text).toBe("reply");
-    expect(sendRequest.durability).toBe("best_effort");
-    expect(sendRequest.replyToMode).toBe("first");
-    expect(sendRequest.threadId).toBe(777);
-    expect(sendRequest.session).toEqual({
-      key: "agent:main:test:peer",
-      agentId: "main",
-      requesterAccountId: "acct",
-      requesterSenderId: "sender-1",
-      conversationType: "group",
-      conversationKind: "group",
-    });
-    expect(resolveOutboundDurableFinalDeliverySupport).toHaveBeenCalledTimes(1);
-    const supportRequest = latestDurableSupportRequest();
-    expect(supportRequest.channel).toBe("telegram");
-    expect(supportRequest.requirements).toEqual({
-      text: true,
-      thread: true,
-      messageSendingHooks: true,
-    });
-  });
-
-  it("returns durable delivery result to the buffered dispatcher", async () => {
-    sendDurableMessageBatch.mockResolvedValueOnce(createDurableSendResult(["tg-1", "tg-2"]));
-    let deliveredResult: unknown;
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(
-      async (params: Parameters<DispatchReplyWithBufferedBlockDispatcher>[0]) => {
-        deliveredResult = await params.dispatcherOptions.deliver(
-          { text: "reply" },
-          { kind: "final" },
-        );
-        return {
-          queuedFinal: true,
-          counts: { tool: 0, block: 0, final: 1 },
-        };
-      },
-    ) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "telegram",
-      accountId: "acct",
-      routeSessionKey: "agent:main:telegram:peer",
-      ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver: vi.fn(), durable: { replyToMode: "first" } },
-    });
-
-    const delivered = deliveryResult(deliveredResult);
-    expect(delivered.messageIds).toEqual(["tg-1", "tg-2"]);
-    expect(delivered.receipt?.platformMessageIds).toEqual(["tg-1", "tg-2"]);
-    expect(delivered.visibleReplySent).toBe(true);
-  });
-
-  it("maps durable hook cancellation to typed routed suppression", async () => {
-    sendDurableMessageBatch.mockResolvedValueOnce({
-      status: "suppressed",
-      results: [],
-      receipt: { platformMessageIds: [], parts: [], sentAt: 1 },
-      reason: "cancelled_by_message_sending_hook",
-      payloadOutcomes: [
-        {
-          index: 0,
-          status: "suppressed",
-          reason: "cancelled_by_message_sending_hook",
-          hookEffect: { cancelReason: "policy", metadata: { source: "test" } },
-        },
-      ],
-    });
-    const onDelivered = vi.fn();
-
-    const result = await dispatchRoutedChannelTurn({
-      cfg,
-      channel: "telegram",
-      route: { agentId: "main", sessionKey: "agent:main:telegram:peer" },
-      ctxPayload: createCtx({ Surface: "telegram", To: "chat-1" }),
-      delivery: {
-        deliver: vi.fn(),
-        durable: { replyToMode: "first" },
-        onDelivered,
-      },
-    });
-
-    expect(onDelivered).toHaveBeenCalledWith(
-      { text: "reply" },
-      { kind: "final" },
-      expect.objectContaining({
-        visibleReplySent: false,
-        suppression: {
-          reason: "cancelled_by_message_sending_hook",
-          cancelReason: "policy",
-          metadata: { source: "test" },
-        },
-      }),
-    );
-    expectDispatched(result);
-    expect(result.dispatchResult).toMatchObject({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-    });
-  });
-
-  it("prepares payloads before durable enqueue and observes handled delivery", async () => {
-    sendDurableMessageBatch.mockResolvedValueOnce(createDurableSendResult(["tlon-1"]));
-    const onDelivered = vi.fn();
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await dispatchTestAssembledTurn({
-      channel: "tlon",
-      accountId: "acct",
-      routeSessionKey: "agent:main:tlon:peer",
-      ctxPayload: createCtx({ To: "chat/~nec/general", OriginatingTo: "chat/~nec/general" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        deliver: vi.fn(),
-        durable: (payload) => ({
-          replyToMode: "first",
-          requiredCapabilities: { text: payload.text?.includes("Generated") === true },
-        }),
-        preparePayload: (payload) => ({
-          ...payload,
-          text: `${payload.text}\n\n_[Generated by test]_`,
-        }),
-        observeMessageSent: true,
-        onDelivered,
-      },
-    });
-
-    expect(sendDurableMessageBatch).toHaveBeenCalledTimes(1);
-    expect(latestDurableSendRequest().payloads?.[0]?.text).toBe("reply\n\n_[Generated by test]_");
-    expect(resolveOutboundDurableFinalDeliverySupport).toHaveBeenCalledTimes(1);
-    expect(latestDurableSupportRequest().requirements).toEqual({
-      text: true,
-    });
-    expect(onDelivered).toHaveBeenCalledTimes(1);
-    const [deliveredPayload, deliveredInfo, deliveredResult] = onDelivered.mock
-      .calls[0] as unknown as [ReplyPayload, unknown, DeliveryResult];
-    expect(deliveredPayload.text).toBe("reply\n\n_[Generated by test]_");
-    expect(deliveredInfo).toEqual({ kind: "final" });
-    expect(deliveredResult.visibleReplySent).toBe(true);
-    // The durable outbound pipeline owns message_sent; the turn lifecycle must not duplicate it.
-    expect(emitMessageSent).not.toHaveBeenCalled();
-  });
-
-  it("falls back before queueing when durable outbound delivery is unsupported", async () => {
-    resolveOutboundDurableFinalDeliverySupport.mockResolvedValueOnce({
-      ok: false,
-      reason: "missing_outbound_handler",
-    });
-    const deliver = vi.fn(async () => ({ messageIds: ["legacy-1"], visibleReplySent: true }));
-    let deliveredResult: unknown;
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(
-      async (params: Parameters<DispatchReplyWithBufferedBlockDispatcher>[0]) => {
-        deliveredResult = await params.dispatcherOptions.deliver(
-          { text: "reply" },
-          { kind: "final" },
-        );
-        return {
-          queuedFinal: true,
-          counts: { tool: 0, block: 0, final: 1 },
-        };
-      },
-    ) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "telegram",
-      accountId: "acct",
-      routeSessionKey: "agent:main:telegram:peer",
-      ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver, durable: { replyToMode: "first" } },
-    });
-
-    expect(resolveOutboundDurableFinalDeliverySupport).toHaveBeenCalledTimes(1);
-    const supportRequest = latestDurableSupportRequest();
-    expect(supportRequest.channel).toBe("telegram");
-    expect(supportRequest.requirements).toEqual({
-      text: true,
-      messageSendingHooks: true,
-    });
-    expect(deliverOutboundPayloads).not.toHaveBeenCalled();
-    expect(deliver).toHaveBeenCalledWith({ text: "reply" }, { kind: "final" });
-    const delivered = deliveryResult(deliveredResult);
-    expect(delivered.messageIds).toEqual(["legacy-1"]);
-    expect(delivered.visibleReplySent).toBe(true);
-  });
-
-  it("treats durable outbound support preflight failures as terminal", async () => {
-    resolveOutboundDurableFinalDeliverySupport.mockRejectedValueOnce(new Error("preflight failed"));
-    const deliver = vi.fn(async () => ({ messageIds: ["legacy-1"], visibleReplySent: true }));
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "telegram",
-        accountId: "acct",
-        routeSessionKey: "agent:main:telegram:peer",
-        ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: { deliver, durable: { replyToMode: "first" } },
-      }),
-    ).rejects.toThrow("preflight failed");
-
-    expect(deliverOutboundPayloads).not.toHaveBeenCalled();
-    expect(deliver).not.toHaveBeenCalled();
-  });
-
-  it("preserves durable partial-send visibility when generic delivery throws", async () => {
-    const error = new Error("second chunk failed");
-    sendDurableMessageBatch.mockResolvedValueOnce({
-      status: "partial_failed",
-      results: [{ channel: "telegram", messageId: "tg-1" }],
-      receipt: {
-        primaryPlatformMessageId: "tg-1",
-        platformMessageIds: ["tg-1"],
-        parts: [{ platformMessageId: "tg-1", kind: "text", index: 0 }],
-        sentAt: 1,
-      },
-      error,
-      sentBeforeError: true,
-    });
-    const deliver = vi.fn(async () => ({ messageIds: ["legacy-1"], visibleReplySent: true }));
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "telegram",
-        accountId: "acct",
-        routeSessionKey: "agent:main:telegram:peer",
-        ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: { deliver, durable: { replyToMode: "first" } },
-      }),
-    ).rejects.toMatchObject({
-      sentBeforeError: true,
-      visibleReplySent: true,
-    });
-
-    expect(deliver).not.toHaveBeenCalled();
-  });
-
-  it("preserves visible delivery when post-delivery observers throw", async () => {
-    const error = new Error("observer failed");
-    const deliver = vi.fn(async () => ({ messageIds: ["local-1"], visibleReplySent: true }));
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "telegram",
-        accountId: "acct",
-        routeSessionKey: "agent:main:telegram:peer",
-        ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: {
-          deliver,
-          durable: false,
-          onDelivered: () => {
-            throw error;
-          },
-        },
-      }),
-    ).rejects.toMatchObject({
-      sentBeforeError: true,
-      visibleReplySent: true,
-    });
-    expect(error).toMatchObject({
-      sentBeforeError: true,
-      visibleReplySent: true,
-    });
-  });
-
-  it("returns custom delivery result to the buffered dispatcher", async () => {
-    let deliveredResult: unknown;
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(
-      async (params: Parameters<DispatchReplyWithBufferedBlockDispatcher>[0]) => {
-        deliveredResult = await params.dispatcherOptions.deliver(
-          { text: "reply" },
-          { kind: "final" },
-        );
-        return {
-          queuedFinal: true,
-          counts: { tool: 0, block: 0, final: 1 },
-        };
-      },
-    ) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        durable: false,
-        deliver: vi.fn(async () => ({ messageIds: ["local-1"], visibleReplySent: true })),
-      },
-    });
-
-    const delivered = deliveryResult(deliveredResult);
-    expect(delivered.messageIds).toEqual(["local-1"]);
-    expect(delivered.visibleReplySent).toBe(true);
-  });
-
-  it("observes provider-finalized content and identity after deferred delivery settles", async () => {
-    const events: string[] = [];
-    emitMessageSent.mockImplementation((event) => {
-      events.push("message_sent");
-      return event;
-    });
-    let resolveFinalization!: (result: {
-      content: string;
-      messageIds: string[];
-      visibleReplySent: true;
-    }) => void;
-    const finalization = new Promise<{
-      content: string;
-      messageIds: string[];
-      visibleReplySent: true;
-    }>((resolve) => {
-      resolveFinalization = resolve;
-    });
-    const deliver = vi.fn(async () => {
-      events.push("deliver");
-      return { visibleReplySent: false, finalization };
-    });
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      params.replyOptions?.onAgentRunStart?.("run-finalized");
-      await params.dispatcherOptions.deliver({ text: "pre-final text" }, { kind: "final" });
-      events.push("provider-finalized");
-      resolveFinalization({
-        content: "provider final text",
-        messageIds: ["om-final"],
-        visibleReplySent: true,
-      });
-      return { queuedFinal: true, counts: { tool: 0, block: 0, final: 1 } };
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "feishu",
-      accountId: "acct",
-      routeSessionKey: "agent:main:feishu:peer",
-      ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu", OriginatingTo: "oc_chat" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver, observeMessageSent: true },
-    });
-
-    expect(events).toEqual(["deliver", "provider-finalized", "message_sent"]);
-    expect(emitMessageSent).toHaveBeenCalledOnce();
-    expect(emitMessageSent).toHaveBeenCalledWith({
-      success: true,
-      content: "provider final text",
-      messageId: "om-final",
-    });
-    expect(createMessageSentEmitter).toHaveBeenCalledWith(
-      expect.objectContaining({
-        channel: "feishu",
-        to: "oc_chat",
-        runId: "run-finalized",
-        sessionKeyForInternalHooks: "agent:main:feishu:peer",
-      }),
-    );
-  });
-
-  it("emits and observes ordinary delivery before buffered dispatch continues", async () => {
-    const events: string[] = [];
-    emitMessageSent.mockImplementation(() => {
-      events.push("message_sent");
-    });
-    const onDelivered = vi.fn(() => {
-      events.push("onDelivered");
-    });
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      await params.dispatcherOptions.deliver({ text: "ordinary" }, { kind: "final" });
-      events.push("after-deliver");
-      return { queuedFinal: true, counts: { tool: 0, block: 0, final: 1 } };
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "feishu",
-      routeSessionKey: "agent:main:feishu:peer",
-      ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        observeMessageSent: true,
-        deliver: async () => ({
-          content: "ordinary",
-          messageIds: ["om-ordinary"],
-          visibleReplySent: true,
-        }),
-        onDelivered,
-      },
-    });
-
-    expect(events).toEqual(["message_sent", "onDelivered", "after-deliver"]);
-    expect(onDelivered).toHaveBeenCalledOnce();
-  });
-
-  it("does not emit a second failure when a post-send observer throws", async () => {
-    const observerError = new Error("observer failed");
-    const onError = vi.fn();
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "feishu",
-        routeSessionKey: "agent:main:feishu:peer",
-        ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher: createDispatch(),
-        delivery: {
-          observeMessageSent: true,
-          deliver: async () => ({ messageIds: ["om-visible"], visibleReplySent: true }),
-          onDelivered: () => {
-            throw observerError;
-          },
-          onError,
-        },
-      }),
-    ).rejects.toBe(observerError);
-
-    expect(emitMessageSent).toHaveBeenCalledOnce();
-    expect(emitMessageSent).toHaveBeenCalledWith({
-      success: true,
-      content: "reply",
-      messageId: "om-visible",
-    });
-    expect(onError).not.toHaveBeenCalled();
-  });
-
-  it("observes early finalization rejection before reporting partial delivery", async () => {
-    let rejectFinalization!: (error: unknown) => void;
-    const finalization = new Promise<{
-      content: string;
-      messageIds: string[];
-      visibleReplySent: true;
-    }>((_resolve, reject) => {
-      rejectFinalization = reject;
-    });
-    const catchSpy = vi.spyOn(finalization, "catch");
-    const partialError = Object.assign(
-      new Error("final edit failed", { cause: new Error("provider rejected edit") }),
-      {
-        code: "CHANNEL_PARTIAL_DELIVERY",
-        deliveryResult: {
-          content: "accepted preview",
-          messageIds: ["om-preview"],
-          visibleReplySent: true,
-        },
-      },
-    );
-    const onError = vi.fn();
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      await params.dispatcherOptions.deliver({ text: "requested final" }, { kind: "final" });
-      expect(catchSpy).toHaveBeenCalledOnce();
-      rejectFinalization(partialError);
-      await new Promise<void>((resolve) => {
-        setImmediate(resolve);
-      });
-      return { queuedFinal: true, counts: { tool: 0, block: 0, final: 1 } };
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "feishu",
-        routeSessionKey: "agent:main:feishu:peer",
-        ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: {
-          observeMessageSent: true,
-          deliver: async () => ({ visibleReplySent: false, finalization }),
-          onError,
-        },
-      }),
-    ).rejects.toBe(partialError);
-
-    expect(emitMessageSent).toHaveBeenCalledOnce();
-    expect(emitMessageSent).toHaveBeenCalledWith({
-      success: false,
-      content: "accepted preview",
-      error: "final edit failed | provider rejected edit",
-      messageId: "om-preview",
-    });
-    expect(onError).toHaveBeenCalledOnce();
-    expect(onError).toHaveBeenCalledWith(partialError, { kind: "final" });
-  });
-
-  it("preserves deferred partial delivery when dispatch also fails", async () => {
-    let rejectFinalization!: (error: unknown) => void;
-    const finalization = new Promise<{
-      content: string;
-      messageIds: string[];
-      visibleReplySent: true;
-    }>((_resolve, reject) => {
-      rejectFinalization = reject;
-    });
-    const dispatchError = new Error("stream close failed");
-    const settlementError = Object.assign(new Error("static fallback failed"), {
-      code: "CHANNEL_PARTIAL_DELIVERY",
-      deliveryResult: {
-        content: "accepted preview",
-        messageIds: ["om-preview"],
-        visibleReplySent: true,
-      },
-    });
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      await params.dispatcherOptions.deliver({ text: "requested final" }, { kind: "final" });
-      rejectFinalization(settlementError);
-      throw dispatchError;
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "feishu",
-        routeSessionKey: "agent:main:feishu:peer",
-        ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: {
-          observeMessageSent: true,
-          deliver: async () => ({ visibleReplySent: false, finalization }),
-        },
-      }),
-    ).rejects.toBe(settlementError);
-
-    expect(emitMessageSent).toHaveBeenCalledOnce();
-    expect(emitMessageSent).toHaveBeenCalledWith({
-      success: false,
-      content: "accepted preview",
-      error: "static fallback failed",
-      messageId: "om-preview",
-    });
-  });
-
-  it("prefers a later visible partial error across deferred payloads", async () => {
-    let rejectFirst!: (error: unknown) => void;
-    let rejectSecond!: (error: unknown) => void;
-    const firstFinalization = new Promise<never>((_resolve, reject) => {
-      rejectFirst = reject;
-    });
-    const secondFinalization = new Promise<never>((_resolve, reject) => {
-      rejectSecond = reject;
-    });
-    const firstError = new Error("first finalization failed");
-    const partialError = Object.assign(new Error("second finalization failed"), {
-      code: "CHANNEL_PARTIAL_DELIVERY",
-      deliveryResult: {
-        content: "accepted second preview",
-        messageIds: ["om-second-preview"],
-        visibleReplySent: true,
-      },
-    });
-    const deliver = vi
-      .fn()
-      .mockResolvedValueOnce({ visibleReplySent: false, finalization: firstFinalization })
-      .mockResolvedValueOnce({ visibleReplySent: false, finalization: secondFinalization });
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      await params.dispatcherOptions.deliver({ text: "first requested" }, { kind: "final" });
-      await params.dispatcherOptions.deliver({ text: "second requested" }, { kind: "final" });
-      rejectFirst(firstError);
-      rejectSecond(partialError);
-      return { queuedFinal: true, counts: { tool: 0, block: 0, final: 2 } };
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "feishu",
-        routeSessionKey: "agent:main:feishu:peer",
-        ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: { observeMessageSent: true, deliver },
-      }),
-    ).rejects.toBe(partialError);
-
-    expect(emitMessageSent).toHaveBeenCalledTimes(2);
-    expect(emitMessageSent).toHaveBeenNthCalledWith(1, {
-      success: false,
-      content: "first requested",
-      error: "first finalization failed",
-      messageId: undefined,
-    });
-    expect(emitMessageSent).toHaveBeenNthCalledWith(2, {
-      success: false,
-      content: "accepted second preview",
-      error: "second finalization failed",
-      messageId: "om-second-preview",
-    });
-  });
-
-  it("suppresses message_sent when the adapter proves provider dispatch never began", async () => {
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async (params) => {
-      try {
-        await params.dispatcherOptions.deliver({ text: "reply" }, { kind: "final" });
-      } catch {
-        // The buffered dispatcher already owns delivery-error reporting.
-      }
-      return { queuedFinal: false, counts: { tool: 0, block: 0, final: 0 } };
-    }) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "feishu",
-      routeSessionKey: "agent:main:feishu:peer",
-      ctxPayload: createCtx({ Surface: "feishu", Provider: "feishu" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        observeMessageSent: true,
-        deliver: async () => {
-          throw new PlatformMessageNotDispatchedError("local media load failed", {
-            cause: new Error("missing file"),
-          });
-        },
-      },
-    });
-
-    expect(emitMessageSent).not.toHaveBeenCalled();
-  });
-
-  it("does not use durable outbound delivery when durable options are omitted", async () => {
-    const deliver = vi.fn(async () => ({ messageIds: ["local-1"], visibleReplySent: true }));
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await dispatchTestAssembledTurn({
-      channel: "telegram",
-      accountId: "acct",
-      routeSessionKey: "agent:main:telegram:peer",
-      ctxPayload: createCtx({ To: "123", OriginatingTo: "123" }),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver },
-    });
-
-    expect(deliverOutboundPayloads).not.toHaveBeenCalled();
-    expect(deliver).toHaveBeenCalledWith({ text: "reply" }, { kind: "final" });
-  });
-
-  it("prepares payloads and observes legacy delivery results", async () => {
-    const onDelivered = vi.fn();
-    const deliver = vi.fn(async () => ({ messageIds: ["local-1"], visibleReplySent: true }));
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await dispatchTestAssembledTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        deliver,
-        preparePayload: (payload) => ({ ...payload, text: `${payload.text}!` }),
-        onDelivered,
-      },
-    });
-
-    expect(deliver).toHaveBeenCalledWith({ text: "reply!" }, { kind: "final" });
-    expect(onDelivered).toHaveBeenCalledTimes(1);
-    const [deliveredPayload, deliveredInfo, deliveredResult] = onDelivered.mock
-      .calls[0] as unknown as [ReplyPayload, unknown, DeliveryResult];
-    expect(deliveredPayload).toEqual({ text: "reply!" });
-    expect(deliveredInfo).toEqual({ kind: "final" });
-    expect(deliveredResult.messageIds).toEqual(["local-1"]);
-    expect(deliveredResult.visibleReplySent).toBe(true);
-  });
-
-  it("assembles channel message reply pipeline options inside the turn kernel", async () => {
-    const deliver = vi.fn(async () => ({ messageIds: ["local-1"], visibleReplySent: true }));
-    const transformReplyPayload = vi.fn((payload: ReplyPayload) => ({
-      ...payload,
-      text: `${payload.text} from pipeline`,
-    }));
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(
-      async (params: Parameters<DispatchReplyWithBufferedBlockDispatcher>[0]) => {
-        const transformed = params.dispatcherOptions.transformReplyPayload?.({ text: "reply" });
-        await params.dispatcherOptions.deliver(transformed ?? { text: "missing" }, {
-          kind: "final",
-        });
-        return {
-          queuedFinal: true,
-          counts: { tool: 0, block: 0, final: 1 },
-        };
-      },
-    ) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver },
-      replyPipeline: { transformReplyPayload },
-    });
-
-    expect(transformReplyPayload).toHaveBeenCalledWith({ text: "reply" });
-    expect(deliver).toHaveBeenCalledWith({ text: "reply from pipeline" }, { kind: "final" });
-  });
-
-  it("records inbound session before dispatching delivery", async () => {
-    const events: string[] = [];
-    const deliver = vi.fn(async () => {
-      events.push("deliver");
-    });
-    const recordInboundSession = createRecordInboundSession(events);
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch(events);
-
-    const result = await dispatchTestAssembledTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver },
-      record: {
-        onRecordError: vi.fn(),
-      },
-    });
-
-    expectDispatched(result);
-    expect(result.dispatchResult?.counts.final).toBe(1);
-    expect(events).toEqual(["record", "dispatch", "deliver"]);
-    expect(recordInboundSession).toHaveBeenCalledTimes(1);
-    const [recordRequest] = (recordInboundSession as unknown as ReturnType<typeof vi.fn>).mock
-      .calls[0] as unknown as [{ sessionKey?: string; storePath?: string }];
-    expect(recordRequest.sessionKey).toBe("agent:main:test:peer");
-    expect(recordRequest.storePath).toBe("/tmp/sessions.json");
-    expect(deliver).toHaveBeenCalledWith({ text: "reply" }, { kind: "final" });
-  });
-
-  it("can record a target session without changing the command dispatch session", async () => {
-    const log = vi.fn();
-    const recordInboundSession = createRecordInboundSession();
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-    const commandSessionKey = "agent:main:command:telegram:42";
-    const targetSessionKey = "agent:main:telegram:group:42:topic:7";
-
-    const result = await dispatchTestAssembledTurn({
-      channel: "telegram",
-      routeSessionKey: commandSessionKey,
-      ctxPayload: createCtx({
-        AgentId: "main",
-        SessionKey: commandSessionKey,
-        CommandTargetSessionKey: targetSessionKey,
-        SessionTranscriptContext: { historyLimit: 1 },
-      }),
-      recordInboundSession,
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver: async () => ({ visibleReplySent: true }) },
-      record: { sessionKey: targetSessionKey },
-      log,
-    });
-
-    expectDispatched(result);
-    expect(result.routeSessionKey).toBe(commandSessionKey);
-    expect(result.ctxPayload.SessionKey).toBe(commandSessionKey);
-    expect(recordInboundSession).toHaveBeenCalledWith(
-      expect.objectContaining({ sessionKey: targetSessionKey }),
-    );
-    expect(readRecentUserAssistantTextForSession).toHaveBeenCalledWith(
-      expect.objectContaining({
-        agentId: "main",
-        sessionKey: targetSessionKey,
-        storePath: "/tmp/sessions.json",
-      }),
-    );
-    const recordEvents = log.mock.calls
-      .map(([event]) => event as TurnLogEvent)
-      .filter((event) => event.stage === "record");
-    expect(recordEvents).toEqual([
-      expect.objectContaining({ event: "start", sessionKey: targetSessionKey }),
-      expect.objectContaining({ event: "done", sessionKey: targetSessionKey }),
-    ]);
-    expect(dispatchReplyWithBufferedBlockDispatcher).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ctx: expect.objectContaining({ SessionKey: commandSessionKey }),
-      }),
-    );
-  });
-
-  it("rejects an empty explicit record session before recording or dispatch", async () => {
-    const recordInboundSession = createRecordInboundSession();
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "telegram",
-        routeSessionKey: "agent:main:command:telegram:42",
-        ctxPayload: createCtx(),
-        recordInboundSession,
-        dispatchReplyWithBufferedBlockDispatcher,
-        delivery: { deliver: async () => ({ visibleReplySent: true }) },
-        record: { sessionKey: "  " },
-      }),
-    ).rejects.toThrow("Channel turn record.sessionKey must be non-empty.");
-    expect(recordInboundSession).not.toHaveBeenCalled();
-    expect(dispatchReplyWithBufferedBlockDispatcher).not.toHaveBeenCalled();
-  });
-
-  it("rejects surrounding whitespace in an explicit record session", async () => {
-    await expect(
-      dispatchTestAssembledTurn({
-        channel: "telegram",
-        routeSessionKey: "agent:main:command:telegram:42",
-        ctxPayload: createCtx(),
-        recordInboundSession: createRecordInboundSession(),
-        dispatchReplyWithBufferedBlockDispatcher: createDispatch(),
-        delivery: { deliver: async () => ({ visibleReplySent: true }) },
-        record: { sessionKey: " agent:main:telegram:group:42 " },
-      }),
-    ).rejects.toThrow("Channel turn record.sessionKey must not include surrounding whitespace.");
-  });
-
-  it("runs prepared dispatches after recording session metadata", async () => {
-    const events: string[] = [];
-    const log = vi.fn();
-    const recordInboundSession = createRecordInboundSession(events);
-    const runDispatch = vi.fn(async () => {
-      events.push("dispatch");
-      return {
-        queuedFinal: true,
-        counts: { tool: 0, block: 0, final: 1 },
-      };
-    });
-
-    const result = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      log,
-      messageId: "msg-1",
-      record: {
-        onRecordError: vi.fn(),
-      },
-    });
-
-    expect(events).toEqual(["record", "dispatch"]);
-    expectDispatched(result);
-    expect(result.dispatchResult?.queuedFinal).toBe(true);
-    expect(loggedEvents(log)).toEqual([
-      { stage: "record", event: "start", messageId: "msg-1" },
-      { stage: "record", event: "done", messageId: "msg-1" },
-      { stage: "dispatch", event: "start", messageId: "msg-1" },
-      { stage: "dispatch", event: "done", messageId: "msg-1" },
-    ]);
-  });
-
-  it("keeps channel message, harness, usage, and model diagnostics in one trace scope", async () => {
-    const diagnostics: DiagnosticEventPayload[] = [];
-    const unsubscribe = onInternalDiagnosticEvent((event) => {
-      if (
-        event.type === "message.processed" ||
-        event.type === "harness.run.started" ||
-        event.type === "model.usage" ||
-        event.type === "model.call.started" ||
-        event.type === "log.record"
-      ) {
-        diagnostics.push(event);
-      }
-    });
-    const recordInboundSession = createRecordInboundSession();
-    const runDispatch = vi.fn(async () => {
-      const messageTrace = getActiveDiagnosticTraceContext();
-      if (!messageTrace) {
-        throw new Error("expected active channel message trace");
-      }
-      const harnessTrace = freezeDiagnosticTraceContext(
-        createChildDiagnosticTraceContext(messageTrace),
-      );
-      const runTrace = freezeDiagnosticTraceContext(
-        createChildDiagnosticTraceContext(harnessTrace),
-      );
-      const modelCallTrace = freezeDiagnosticTraceContext(
-        createChildDiagnosticTraceContext(runTrace),
-      );
-      const usageTrace = freezeDiagnosticTraceContext(
-        createChildDiagnosticTraceContext(harnessTrace),
-      );
-      getChildLogger({ subsystem: "diagnostic" }).info({ runId: "run-1" }, "channel lifecycle log");
-
-      emitTrustedDiagnosticEvent({
-        type: "harness.run.started",
-        runId: "run-1",
-        harnessId: "codex",
-        pluginId: "codex",
-        provider: "openai",
-        model: "gpt-5.5",
-        channel: "slack",
-        trace: harnessTrace,
-      });
-      emitTrustedDiagnosticEvent({
-        type: "model.call.started",
-        runId: "run-1",
-        callId: "call-1",
-        provider: "openai",
-        model: "gpt-5.5",
-        api: "openai-codex-responses",
-        transport: "stdio",
-        trace: modelCallTrace,
-      });
-      emitTrustedDiagnosticEvent({
-        type: "model.usage",
-        sessionKey: "agent:main:slack:channel:c1",
-        channel: "slack",
-        agentId: "main",
-        provider: "openai",
-        model: "gpt-5.5",
-        usage: { input: 10, output: 5, total: 15 },
-        durationMs: 25,
-        trace: usageTrace,
-      });
-      logMessageProcessed({
-        channel: "slack",
-        messageId: "msg-1",
-        chatId: "c1",
-        sessionKey: "agent:main:slack:channel:c1",
-        durationMs: 50,
-        outcome: "completed",
-      });
-      return {
-        queuedFinal: true,
-        counts: { tool: 0, block: 0, final: 1 },
-      };
-    });
-
-    try {
-      await runPreparedChannelTurn({
-        channel: "slack",
-        routeSessionKey: "agent:main:slack:channel:c1",
-        storePath: "/tmp/sessions.json",
-        ctxPayload: createCtx({ SessionKey: "agent:main:slack:channel:c1" }),
-        recordInboundSession,
-        runDispatch,
-        messageId: "msg-1",
-      });
-      await waitForDiagnosticEventsDrained();
-    } finally {
-      unsubscribe();
-    }
-
-    const message = diagnostics.find((event) => event.type === "message.processed");
-    const harness = diagnostics.find((event) => event.type === "harness.run.started");
-    const usage = diagnostics.find((event) => event.type === "model.usage");
-    const modelCall = diagnostics.find((event) => event.type === "model.call.started");
-    const logRecord = diagnostics.find(
-      (event) => event.type === "log.record" && event.message === "channel lifecycle log",
-    );
-    const traceId = message?.trace?.traceId;
-
-    expect(traceId).toBeTruthy();
-    expect(harness?.trace?.traceId).toBe(traceId);
-    expect(usage?.trace?.traceId).toBe(traceId);
-    expect(modelCall?.trace?.traceId).toBe(traceId);
-    expect(harness?.trace?.parentSpanId).toBe(message?.trace?.spanId);
-    expect(usage?.trace?.parentSpanId).toBe(harness?.trace?.spanId);
-    expect(modelCall?.trace?.parentSpanId).toBeTruthy();
-    expect(modelCall?.trace?.parentSpanId).not.toBe(message?.trace?.spanId);
-    expect(logRecord?.trace?.traceId).toBe(traceId);
-  });
-
-  it("logs a warning when a visible prepared dispatch queues no payloads", async () => {
-    const events: string[] = [];
-    const log = vi.fn();
-    const recordInboundSession = createRecordInboundSession(events);
-    const runDispatch = vi.fn(async () => ({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-    }));
-
-    const result = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      log,
-      messageId: "msg-zero",
-      record: {
-        onRecordError: vi.fn(),
-      },
-    });
-
-    expectDispatched(result);
-    expect(result.dispatchResult?.queuedFinal).toBe(false);
-    expect(log.mock.calls).toContainEqual([
-      expect.objectContaining({
-        stage: "dispatch",
-        event: "warning",
-        messageId: "msg-zero",
-        reason: "zero-count-visible-dispatch",
-      }),
-    ]);
-  });
-
-  it("does not warn for observed-path deliveries with zero queued counts", async () => {
-    const events: string[] = [];
-    const log = vi.fn();
-    const recordInboundSession = createRecordInboundSession(events);
-    // Observed-delivery path: queuedFinal false and all counts zero, but the reply was
-    // delivered via observedReplyDelivery and must not trip the silent-drop sentinel.
-    const runDispatch = vi.fn(async () => ({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-      observedReplyDelivery: true,
-    }));
-
-    const result = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      log,
-      messageId: "msg-observed",
-      record: {
-        onRecordError: vi.fn(),
-      },
-    });
-
-    expectDispatched(result);
-    expect(result.dispatchResult?.observedReplyDelivery).toBe(true);
-    expect(log.mock.calls).not.toContainEqual([
-      expect.objectContaining({ reason: "zero-count-visible-dispatch" }),
-    ]);
-  });
-
-  it("still warns when a visible turn has zero counts and no observed delivery", async () => {
-    const events: string[] = [];
-    const log = vi.fn();
-    const recordInboundSession = createRecordInboundSession(events);
-    // Guard against over-suppression: a genuinely empty visible dispatch must still warn.
-    const runDispatch = vi.fn(async () => ({
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-      observedReplyDelivery: false,
-    }));
-
-    const result = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      log,
-      messageId: "msg-empty",
-      record: {
-        onRecordError: vi.fn(),
-      },
-    });
-
-    expectDispatched(result);
-    expect(result.dispatchResult?.observedReplyDelivery).toBe(false);
-    expect(log.mock.calls).toContainEqual([
-      expect.objectContaining({
-        stage: "dispatch",
-        event: "warning",
-        messageId: "msg-empty",
-        reason: "zero-count-visible-dispatch",
-      }),
-    ]);
-  });
-
-  it("drops direct prepared turns with bot-loop protection before record and dispatch", async () => {
-    const events: string[] = [];
-    const log = vi.fn();
-    const historyMap = new Map<string, HistoryEntry[]>([
-      ["room", [{ sender: "User", body: "queued before suppression" }]],
-    ]);
-    const recordInboundSession = createRecordInboundSession(events);
-    const runDispatch = vi.fn(async () => {
-      events.push("dispatch");
-      return {
-        queuedFinal: true,
-        counts: { tool: 0, block: 0, final: 1 },
-      };
-    });
-    const onDispatchSkipped = vi.fn();
-    const botLoopProtection = {
-      scopeId: "prepared-loop-test",
-      conversationId: "room",
-      senderId: "bot-a",
-      receiverId: "bot-b",
-      config: { maxEventsPerWindow: 1, windowSeconds: 60, cooldownSeconds: 60 },
-      defaultEnabled: true,
-    };
-
-    const first = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      runDispatchLifecycle: {
-        turnAdoptionLifecycle: undefined,
-        onDispatchSkipped,
-      },
-      botLoopProtection: { ...botLoopProtection, nowMs: 1_000 },
-    });
-    const second = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession,
-      runDispatch,
-      runDispatchLifecycle: {
-        turnAdoptionLifecycle: undefined,
-        onDispatchSkipped,
-      },
-      log,
-      messageId: "msg-loop",
-      botLoopProtection: { ...botLoopProtection, nowMs: 1_001 },
-      history: {
-        isGroup: true,
-        historyKey: "room",
-        historyMap,
-        limit: 50,
-      },
-    });
-
-    expect(first.dispatched).toBe(true);
-    expect(second).toMatchObject({
-      admission: { kind: "drop", reason: "bot-loop-protection" },
-      dispatched: false,
-      routeSessionKey: "agent:main:test:peer",
-    });
-    expect(events).toEqual(["record", "dispatch"]);
-    expect(recordInboundSession).toHaveBeenCalledTimes(1);
-    expect(runDispatch).toHaveBeenCalledTimes(1);
-    expect(onDispatchSkipped).toHaveBeenCalledWith("botLoopProtection");
-    expect(historyMap.get("room")).toStrictEqual([]);
-    expect(loggedEvents(log)).toEqual([
-      { stage: "authorize", event: "drop", messageId: "msg-loop" },
-    ]);
-  });
-
-  it("drops a recorded Discord webhook echo after thread unbind before record and dispatch", async () => {
-    const recordInboundSession = createRecordInboundSession();
-    const dispatchReplyWithBufferedBlockDispatcher = createDispatch();
-    const deliver = vi.fn();
-    const onAdopted = vi.fn(async () => {});
-    const log = vi.fn();
-    const historyMap = new Map([["thread-1", [{ sender: "Bot", body: "echo" }]]]);
-    recordOutboundMessageIdentity({
-      channel: "discord",
-      accountId: "default",
-      conversationId: "thread-1",
-      sourceId: "webhook-1",
-    });
-
-    // Unbinding removes Discord's thread route, not the core-owned outbound identity.
-    const result = await dispatchTestAssembledTurn({
-      channel: "discord",
-      accountId: "default",
-      routeSessionKey: "agent:main:discord:channel:thread-1",
-      ctxPayload: createCtx({
-        Provider: "discord",
-        Surface: "discord",
-        ChatId: "thread-1",
-        MessageSid: "webhook-message-1",
-      }),
-      outboundEchoSourceId: "webhook-1",
-      recordInboundSession,
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: { deliver },
-      turnAdoptionLifecycle: { onAdopted },
-      history: {
-        isGroup: true,
-        historyKey: "thread-1",
-        historyMap,
-        limit: 50,
-      },
-      log,
-    });
-
-    expect(result).toMatchObject({
-      admission: { kind: "drop", reason: "outbound-echo" },
-      dispatched: false,
-      routeSessionKey: "agent:main:discord:channel:thread-1",
-    });
-    expect(recordInboundSession).not.toHaveBeenCalled();
-    expect(dispatchReplyWithBufferedBlockDispatcher).not.toHaveBeenCalled();
-    expect(deliver).not.toHaveBeenCalled();
-    expect(onAdopted).toHaveBeenCalledOnce();
-    expect(historyMap.get("thread-1")).toStrictEqual([]);
-    expect(log).toHaveBeenCalledWith(
-      expect.objectContaining({
-        stage: "authorize",
-        event: "drop",
-        reason: "outbound-echo",
-      }),
-    );
-  });
-
-  it("suppresses direct prepared dispatches for observe-only admission", async () => {
-    const events: string[] = [];
-    const recordInboundSession = createRecordInboundSession(events);
-    const runDispatch = vi.fn(async () => {
-      events.push("dispatch");
-      return {
-        queuedFinal: true,
-        counts: { tool: 0, block: 0, final: 1 },
-      };
-    });
-    const observeOnlyDispatchResult = {
-      queuedFinal: false,
-      counts: { tool: 0, block: 0, final: 0 },
-    };
-    const onDispatchSkipped = vi.fn();
-
-    const result = await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:observer:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx({ SessionKey: "agent:observer:test:peer" }),
-      recordInboundSession,
-      runDispatch,
-      runDispatchLifecycle: {
-        turnAdoptionLifecycle: undefined,
-        onDispatchSkipped,
-      },
-      observeOnlyDispatchResult,
-      admission: { kind: "observeOnly", reason: "broadcast-observer" },
-    });
-
-    expect(events).toEqual(["record"]);
-    expect(runDispatch).not.toHaveBeenCalled();
-    expect(onDispatchSkipped).toHaveBeenCalledWith("observeOnly");
-    expect(result.admission).toEqual({ kind: "observeOnly", reason: "broadcast-observer" });
-    expectDispatched(result);
-    expect(result.dispatchResult).toBe(observeOnlyDispatchResult);
-    expect(hasFinalChannelTurnDispatch(result.dispatchResult)).toBe(false);
-  });
-
-  it("uses noop delivery for direct assembled observe-only dispatch", async () => {
-    const events: string[] = [];
-    const deliver = vi.fn(async () => {
-      events.push("deliver");
-      return { visibleReplySent: true };
-    });
-
-    const result = await dispatchAssembledChannelTurn({
-      cfg,
-      channel: "test",
-      agentId: "observer",
-      routeSessionKey: "agent:observer:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx({ SessionKey: "agent:observer:test:peer" }),
-      recordInboundSession: createRecordInboundSession(events),
-      dispatchReplyWithBufferedBlockDispatcher: createDispatch(events),
-      delivery: { deliver },
-      admission: { kind: "observeOnly", reason: "broadcast-observer" },
-    });
-
-    expect(events).toEqual(["record", "dispatch"]);
-    expect(deliver).not.toHaveBeenCalled();
-    expect(result.admission).toEqual({ kind: "observeOnly", reason: "broadcast-observer" });
-    expect(result.dispatched).toBe(true);
-    if (!result.dispatched) {
-      throw new Error("expected dispatch");
-    }
-    expect(resolveChannelTurnDispatchCounts(result.dispatchResult)).toEqual({
-      tool: 0,
-      block: 0,
-      final: 0,
-    });
-  });
-
-  it("clears pending group history after a successful prepared turn", async () => {
-    const historyMap = new Map([["room-1", [{ sender: "User", body: "queued before reply" }]]]);
-
-    await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:group:room-1",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(),
-      runDispatch: vi.fn(async () => ({
-        queuedFinal: false,
-        counts: { tool: 0, block: 0, final: 0 },
-      })),
-      history: {
-        isGroup: true,
-        historyKey: "room-1",
-        historyMap,
-        limit: 50,
-      },
-    });
-
-    expect(historyMap.get("room-1")).toStrictEqual([]);
-  });
-
-  it("cleans up pre-created dispatchers when session recording fails", async () => {
-    const events: string[] = [];
-    const recordError = new Error("session store failed");
-    const log = vi.fn();
-    const recordInboundSession = vi.fn(async () => {
-      events.push("record");
-      throw recordError;
-    }) as unknown as RecordInboundSession;
-    const runDispatch = vi.fn();
-    const onPreDispatchFailure = vi.fn(async () => {
-      events.push("cleanup");
-    });
-
-    await expect(
-      runPreparedChannelTurn({
-        channel: "test",
-        routeSessionKey: "agent:main:test:peer",
-        storePath: "/tmp/sessions.json",
-        ctxPayload: createCtx(),
-        recordInboundSession,
-        onPreDispatchFailure,
-        runDispatch,
-        log,
-        record: {
-          onRecordError: vi.fn(),
-        },
-      }),
-    ).rejects.toThrow(recordError);
-
-    expect(events).toEqual(["record", "cleanup"]);
-    expect(runDispatch).not.toHaveBeenCalled();
-    expect(onPreDispatchFailure).toHaveBeenCalledWith(recordError);
-    expect(loggedEvents(log)).toEqual([
-      { stage: "record", event: "start" },
-      { stage: "record", event: "error" },
-    ]);
-  });
-
-  it("runs afterRecord only after session recording succeeds and before dispatch", async () => {
-    const events: string[] = [];
-    await runPreparedChannelTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      storePath: "/tmp/sessions.json",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(events),
-      afterRecord: vi.fn(async () => {
-        events.push("afterRecord");
-      }),
-      runDispatch: vi.fn(async () => {
-        events.push("dispatch");
-        return { visibleReplySent: true };
-      }),
-    });
-
-    expect(events).toEqual(["record", "afterRecord", "dispatch"]);
-  });
-
-  it("threads turnAdoptionLifecycle into assembled reply options and fires after recovery persist attempt", async () => {
-    const events: string[] = [];
-    const onAdopted = vi.fn(async () => {
-      events.push("adopted");
-    });
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(
-      async (params: Parameters<DispatchReplyWithBufferedBlockDispatcher>[0]) => {
-        events.push("dispatch-start");
-        // Persist attempt completes before adoption (agent-runner contract).
-        events.push("recovery-persist");
-        await params.replyOptions?.turnAdoptionLifecycle?.onAdopted();
-        events.push("settle");
-        return {
-          queuedFinal: true,
-          counts: { tool: 0, block: 0, final: 1 },
-        };
-      },
-    ) as DispatchReplyWithBufferedBlockDispatcher;
-
-    await dispatchTestAssembledTurn({
-      channel: "test",
-      routeSessionKey: "agent:main:test:peer",
-      ctxPayload: createCtx(),
-      recordInboundSession: createRecordInboundSession(events),
-      dispatchReplyWithBufferedBlockDispatcher,
-      delivery: {
-        deliver: vi.fn(async () => undefined),
-      },
-      turnAdoptionLifecycle: { onAdopted },
-    });
-
-    expect(onAdopted).toHaveBeenCalledOnce();
-    expect(events).toEqual(["record", "dispatch-start", "recovery-persist", "adopted", "settle"]);
-    expect(dispatchReplyWithBufferedBlockDispatcher).toHaveBeenCalledWith(
-      expect.objectContaining({
-        replyOptions: expect.objectContaining({
-          turnAdoptionLifecycle: expect.objectContaining({ onAdopted }),
-        }),
-      }),
-    );
-  });
-
-  it("does not run afterRecord when session recording fails", async () => {
-    const recordError = new Error("session store failed");
-    const afterRecord = vi.fn();
-
-    await expect(
-      runPreparedChannelTurn({
-        channel: "test",
-        routeSessionKey: "agent:main:test:peer",
-        storePath: "/tmp/sessions.json",
-        ctxPayload: createCtx(),
-        recordInboundSession: vi.fn(async () => {
-          throw recordError;
-        }) as unknown as RecordInboundSession,
-        afterRecord,
-        runDispatch: vi.fn(),
-      }),
-    ).rejects.toThrow(recordError);
-
-    expect(afterRecord).not.toHaveBeenCalled();
-  });
-
-  it("handles non-turn event classes without dispatch", async () => {
-    const resolveTurn = vi.fn();
-    const result = await runChannelTurn({
-      channel: "test",
-      raw: {},
-      adapter: {
-        ingest: () => ({ id: "evt-1", rawText: "" }),
-        classify: () => ({ kind: "reaction", canStartAgentTurn: false }),
-        resolveTurn,
-      },
-    });
-
-    expect(result.admission).toEqual({ kind: "handled", reason: "event:reaction" });
-    expect(result.dispatched).toBe(false);
-    expect(resolveTurn).not.toHaveBeenCalled();
-  });
-
-  it("stops on preflight admission drops", async () => {
-    const resolveTurn = vi.fn();
-    const result = await runChannelTurn({
-      channel: "test",
-      raw: {},
-      adapter: {
-        ingest: () => ({ id: "msg-1", rawText: "hello" }),
-        preflight: () => ({ kind: "drop", reason: "missing-mention", recordHistory: true }),
-        resolveTurn,
-      },
-    });
-
-    expect(result.admission).toEqual({
-      kind: "drop",
-      reason: "missing-mention",
-      recordHistory: true,
-    });
-    expect(result.dispatched).toBe(false);
-    expect(resolveTurn).not.toHaveBeenCalled();
-  });
-
-  it("records preflight drop history through the turn kernel", async () => {
-    const historyMap = new Map<string, HistoryEntry[]>();
-    const resolveTurn = vi.fn();
-
-    const result = await runChannelTurn({
-      channel: "test",
-      raw: {},
-      adapter: {
-        ingest: () => ({
-          id: "msg-1",
-          timestamp: 1_700_000_000_000,
-          rawText: "<media:image>",
-        }),
-        preflight: () => ({
-          admission: { kind: "drop", reason: "missing-mention", recordHistory: true },
-          message: {
-            bodyForAgent: "<media:image>",
-            senderLabel: "Alice",
-          },
-          history: {
-            key: "room-1",
-            historyMap,
-            limit: 5,
-            mediaLimit: 2,
-          },
-          media: async () => [
-            { path: "/tmp/a.png", contentType: "image/png", kind: "image" },
-            { path: "https://example.com/b.png", contentType: "image/png", kind: "image" },
-          ],
-        }),
-        resolveTurn,
-      },
-    });
-
-    expect(result.admission).toEqual({
-      kind: "drop",
-      reason: "missing-mention",
-      recordHistory: true,
-    });
-    expect(result.dispatched).toBe(false);
-    expect(resolveTurn).not.toHaveBeenCalled();
-    expect(historyMap.get("room-1")).toEqual([
-      {
-        sender: "Alice",
-        body: "<media:image>",
-        timestamp: 1_700_000_000_000,
-        messageId: "msg-1",
-        media: [
-          { path: "/tmp/a.png", contentType: "image/png", kind: "image", messageId: "msg-1" },
-        ],
-      },
-    ]);
-  });
-
-  it("drops repeated bot-pair turns in the core turn kernel before record and dispatch", async () => {
-    const events: string[] = [];
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementation(createDispatch(events));
-    const onFinalize = vi.fn();
-    recordInboundSessionCore.mockImplementation(async () => {
-      events.push("record");
-    });
-    let nowMs = 1_000;
-    const runOne = async (id: string) =>
-      await runChannelTurn({
-        channel: "test",
-        accountId: "acct",
-        raw: { id },
-        adapter: {
-          ingest: () => ({ id, rawText: "hello" }),
-          resolveTurn: () => ({
-            cfg,
-            channel: "test",
-            accountId: "acct",
-            route: { agentId: "main", sessionKey: "agent:main:test:peer" },
-            ctxPayload: createCtx(),
-            botLoopProtection: {
-              scopeId: "acct",
-              conversationId: "room",
-              senderId: "bot-a",
-              receiverId: "bot-b",
-              config: { maxEventsPerWindow: 1, windowSeconds: 60, cooldownSeconds: 60 },
-              defaultEnabled: true,
-              nowMs: nowMs++,
-            },
-            delivery: { deliver: async () => ({ visibleReplySent: true }) },
-          }),
-          onFinalize,
-        },
-      });
-
-    const first = await runOne("msg-1");
-    const second = await runOne("msg-2");
-
-    expect(first.dispatched).toBe(true);
-    expect(second).toEqual({
-      admission: { kind: "drop", reason: "bot-loop-protection" },
-      dispatched: false,
-      ctxPayload: createCtx(),
-      routeSessionKey: "agent:main:test:peer",
-    });
-    expect(events).toEqual(["record", "dispatch"]);
-    expect(onFinalize).toHaveBeenCalledTimes(2);
-    const [, suppressed] = onFinalize.mock.calls;
-    expect(suppressed?.[0]).toMatchObject({
-      admission: { kind: "drop", reason: "bot-loop-protection" },
-      dispatched: false,
-      routeSessionKey: "agent:main:test:peer",
-    });
-  });
-
-  it("runs observe-only preflights through resolve, record, dispatch, and finalize without visible delivery", async () => {
-    const events: string[] = [];
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementation(createDispatch(events));
-    recordInboundSessionCore.mockImplementation(async () => {
-      events.push("record");
-    });
-    const deliver = vi.fn();
-    const onFinalize = vi.fn();
-    const result = await runChannelTurn({
-      channel: "test",
-      raw: {},
-      adapter: {
-        ingest: () => ({ id: "msg-1", rawText: "observe" }),
-        preflight: () => ({ kind: "observeOnly", reason: "broadcast-observer" }),
-        resolveTurn: () => ({
-          cfg,
-          channel: "test",
-          route: {
-            agentId: "observer",
-            dmScope: "per-channel-peer",
-            sessionKey: "agent:observer:test:peer",
-          },
-          ctxPayload: createCtx({ SessionKey: "agent:observer:test:peer" }),
-          delivery: { deliver },
-          record: {
-            onRecordError: vi.fn(),
-          },
-        }),
-        onFinalize,
-      },
-    });
-
-    expect(result.admission).toEqual({
-      kind: "observeOnly",
-      reason: "broadcast-observer",
-    });
-    expect(result.dispatched).toBe(true);
-    expect(events).toEqual(["record", "dispatch"]);
-    expect(dispatchReplyWithRoutedChannelDispatcherCore).toHaveBeenCalledWith(
-      expect.objectContaining({
-        ctx: expect.objectContaining({ DmScope: "per-channel-peer" }),
-        suppressOutboundHooks: true,
-      }),
-    );
-    expect(deliver).not.toHaveBeenCalled();
-    if (!result.dispatched) {
-      throw new Error("expected dispatch");
-    }
-    expect(hasVisibleChannelTurnDispatch(result.dispatchResult)).toBe(false);
-    expect(resolveChannelTurnDispatchCounts(result.dispatchResult)).toEqual({
-      tool: 0,
-      block: 0,
-      final: 0,
-    });
-    expect(onFinalize).toHaveBeenCalledTimes(1);
-    const [finalized] = requireFirstMockCall(onFinalize, "finalize");
-    const finalizedResult = finalizeResult(finalized);
-    expect(finalizedResult.admission).toEqual({
-      kind: "observeOnly",
-      reason: "broadcast-observer",
-    });
-    expect(finalizedResult.dispatched).toBe(true);
-    expect(finalizedResult.routeSessionKey).toBe("agent:observer:test:peer");
-  });
-
-  it("finalizes failed dispatches before rethrowing", async () => {
-    const onFinalize = vi.fn();
-    const dispatchError = new Error("dispatch failed");
-    const dispatchReplyWithBufferedBlockDispatcher = vi.fn(async () => {
-      throw dispatchError;
-    }) as unknown as DispatchReplyWithBufferedBlockDispatcher;
-    dispatchReplyWithRoutedChannelDispatcherCore.mockImplementation(
-      dispatchReplyWithBufferedBlockDispatcher,
-    );
-
-    await expect(
-      runChannelTurn({
-        channel: "test",
-        raw: {},
-        adapter: {
-          ingest: () => ({ id: "msg-1", rawText: "hello" }),
-          resolveTurn: () => ({
-            cfg,
-            channel: "test",
-            route: { agentId: "main", sessionKey: "agent:main:test:peer" },
-            ctxPayload: createCtx(),
-            delivery: { deliver: async () => ({ visibleReplySent: false }) },
-            record: {
-              onRecordError: vi.fn(),
-            },
-          }),
-          onFinalize,
-        },
-      }),
-    ).rejects.toThrow(dispatchError);
-
-    expect(onFinalize).toHaveBeenCalledTimes(1);
-    const [finalized] = requireFirstMockCall(onFinalize, "finalize");
-    const finalizedResult = finalizeResult(finalized);
-    expect(finalizedResult.admission).toEqual({ kind: "dispatch" });
-    expect(finalizedResult.dispatched).toBe(false);
-    expect(finalizedResult.routeSessionKey).toBe("agent:main:test:peer");
-  });
-});
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
